+++ test/src/main.ts
@@ -1,5 +1,6 @@
 import * as scrollAnimation from "../../src/index";
 
+(window as any).scrollAnimation = scrollAnimation;
 // scrollAnimation.add(document.body.querySelector("#test") as HTMLElement);
 // setTimeout(() => {
     scrollAnimation.add(document.body);